Implemented TextContainers for the interactions window.  Changing textcontainer so I can specify key maps (both the filter and its handler), which we'll need for interactions to tell when to run an expression.
Localizing the logic for the Prompt, just so I can understand it.
doing limited cps since instantiating a WeSchemeInteractions will now
be asynchronous.

Needed to CPS more of the interactions code to accomodate the needs
of CodeMirror.

Adjusting options for textcontainer to handle things like readOnly,
and dynamic height.

The interactions text areas watch for balance before evaluation.