Editor Styling Update
NOTE - without being able to test this on a live instance, all of this
was done on a static copy of the editor page (see
www.litech.org/~schanzer/IDE/ ). It's possible this whole thing breaks
when used on a running WeScheme instance, so please double-check before
pushing.

1) move docs/login/console buttons to the subheading, in place of
"Sometimes iPhone…."
2) Move button images into list-style declarations via CSS
3) Remove now-unused SPANs and IMGs from toolbar links
4) Update CSS to create flexible formatting of toolbar buttons, and
accommodate changes to DOM
5) Add DesignRecipe icon